Moved listBoxItem only usage in Select component file
