+++ frontend/app/src/shared/components/aria/select.tsx
@@ -1,15 +1,17 @@
+import { Icon } from "@iconify-icon/react";
 import { ChevronDownIcon } from "lucide-react";
 import {
   Button as AriaButton,
   type ButtonProps as AriaButtonProps,
   ListBox as AriaListBox,
+  ListBoxItem as AriaListBoxItem,
+  type ListBoxItemProps as AriaListBoxItemProps,
   type ListBoxProps as AriaListBoxProps,
   Select as AriaSelect,
   SelectValue as AriaSelectValue,
   composeRenderProps,
 } from "react-aria-components";
 
-import { ListBoxItem } from "@/shared/components/aria/list-box";
 import { focusVisibleStyle } from "@/shared/components/style-rac";
 import { inputStyle } from "@/shared/components/ui/style";
 import { classNames } from "@/shared/utils/common";
@@ -17,8 +19,6 @@ import { classNames } from "@/shared/utils/common";
 import { Popover, type PopoverProps } from "./popover";
 
 export const Select = AriaSelect;
-
-export const SelectItem = ListBoxItem;
 
 export const SelectTrigger = ({ className, children, ...props }: AriaButtonProps) => (
   <AriaButton
@@ -54,3 +54,36 @@ export const SelectList = <T extends object>({ className, ...props }: AriaListBo
     />
   </SelectPopover>
 );
+
+export const SelectItem = <T extends object>({
+  className,
+  children,
+  ...props
+}: AriaListBoxItemProps<T>) => {
+  return (
+    <AriaListBoxItem
+      textValue={props.textValue || (typeof children === "string" ? children : undefined)}
+      className={composeRenderProps(className, (className) =>
+        classNames(
+          "relative flex w-full cursor-default select-none items-center rounded-lg px-2 py-1.5 text-sm outline-hidden",
+          "data-disabled:pointer-events-none data-disabled:opacity-50",
+          "data-focused:bg-gray-100",
+          "data-selection-mode:pl-8",
+          className
+        )
+      )}
+      {...props}
+    >
+      {composeRenderProps(children, (children, renderProps) => (
+        <>
+          {renderProps.isSelected && (
+            <span className="absolute left-2 flex size-4 items-center justify-center">
+              <Icon icon="mdi:check" />
+            </span>
+          )}
+          {children}
+        </>
+      ))}
+    </AriaListBoxItem>
+  );
+};
